Move stage files to better directory to support structure better
