+++ src/renderer/components/FileUploadStages/PickFiles/Stage1.tsx
