+++ src/renderer/components/FileUploadStages/ProcessFiles/Stage3.tsx
@@ -1,5 +1,7 @@
 import * as React from 'react';
 
-export const Stage3 = () => {
+const Stage3 = () => {
     return <div>Stage 3 rendering here</div>
 }
+
+export default Stage3
